✨ Full Preference Submission API Implemented
- Created PreferenceController.java to handle preference submission & retrieval.
- Added PreferenceRepository.java with custom method findByStudentId().
- Organized controllers into epic_2.controller package for clean architecture.
- Moved existing repositories into epic_2.repository package.
- Updated imports and package declarations accordingly.
- Verified database integration working with new API structure.
- Backend now fully supports submitting preferences for students via API.
