--- src/main/java/epic_2/Log.java
@@ -1,24 +1,0 @@
-package epic_2;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "logs")
-@Data
-public class Log {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    private String action;
-
-    private LocalDateTime timestamp;
-}
