--- src/main/java/epic_2/PreferenceRepository.java
@@ -1,6 +1,0 @@
-package epic_2;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PreferenceRepository extends JpaRepository<Preference, Long> {
-}
