+++ src/main/java/epic_2/controller/AdminController.java
@@ -1,5 +1,6 @@
-package epic_2;
+package epic_2.controller;
 
+import epic_2.service.MatchingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
