+++ src/main/java/epic_2/controller/HelloController.java
@@ -1,4 +1,4 @@
-package epic_2;
+package epic_2.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
