+++ src/main/java/epic_2/model/CollegeYear.java
@@ -1,4 +1,4 @@
-package epic_2;
+package epic_2.model;
 
 public enum CollegeYear {
     _1, _2, _3, _4, _5
