+++ src/main/java/epic_2/model/Match.java
@@ -1,4 +1,4 @@
-package epic_2;
+package epic_2.model;
 
 import jakarta.persistence.*;
 import lombok.Data;
@@ -6,9 +6,9 @@ import lombok.Data;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "cvs")
+@Table(name = "matches")
 @Data
-public class CV {
+public class Match {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -22,9 +22,7 @@ public class CV {
     @JoinColumn(name = "residency_id")
     private Residency residency;
 
-    @Column(name = "file_url")
-    private String fileUrl;
+    private Integer round;
 
-    @Column(name = "uploaded_time")
-    private LocalDateTime uploadedTime;
+    private LocalDateTime timestamp;
 }
