+++ src/main/java/epic_2/model/Role.java
@@ -1,4 +1,4 @@
-package epic_2;
+package epic_2.model;
 
 public enum Role {
     student, admin, company
