+++ src/main/java/epic_2/repository/CVRepository.java
@@ -1,5 +1,6 @@
-package epic_2;
+package epic_2.repository;
 
+import epic_2.model.CV;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CVRepository extends JpaRepository<CV, Long> {
