+++ src/main/java/epic_2/repository/InterviewRepository.java
@@ -1,5 +1,6 @@
-package epic_2;
+package epic_2.repository;
 
+import epic_2.model.Interview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InterviewRepository extends JpaRepository<Interview, Long> {
