+++ src/main/java/epic_2/repository/LogRepository.java
@@ -1,5 +1,6 @@
-package epic_2;
+package epic_2.repository;
 
+import epic_2.model.Log;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LogRepository extends JpaRepository<Log, Long> {
