+++ src/main/java/epic_2/repository/PreferenceRepository.java
@@ -1,0 +1,12 @@
+package epic_2.repository;
+
+import epic_2.model.Preference;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface PreferenceRepository extends JpaRepository<Preference, Long> {
+    List<Preference> findByStudentId(Long studentId);
+}
