+++ src/main/java/epic_2/repository/ResidencyRepository.java
@@ -1,5 +1,6 @@
-package epic_2;
+package epic_2.repository;
 
+import epic_2.model.Residency;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ResidencyRepository extends JpaRepository<Residency, Long> {
