+++ src/main/java/epic_2/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package epic_2;
+package epic_2.repository;
 
+import epic_2.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
